unify script
