+++ scripts/publishMaven.js
@@ -1,5 +1,49 @@
 /**
- * Upload artifacts to a nexus staging repo.
+ * Usage:
+ * node script.js -task [upload|promote]
+ * 
+ * upload: Upload artifacts to a nexus staging repo.
+ * promote: Promote a repo to get it picked up by Maven Central.
+ */
+
+const childProcess = require('child_process');
+const fs = require('fs');
+const path = require('path');
+
+const artifactFolder = process.env.artifactFolder;
+const configs = {
+    nexus_ossrhuser: process.env.NEXUS_OSSRHUSER,
+    nexus_ossrhpass: process.env.NEXUS_OSSRHPASS,
+    nexus_stagingProfileId: process.env.NEXUS_STAGINGPROFILEID,
+    gpgpass: process.env.GPGPASS,
+    stagingRepoId: process.env.NEXUS_STAGINGREPOID,
+    groupId: "com.microsoft.java",
+    projectName: "java-debug",
+    releaseVersion: process.env.releaseVersion,    
+    moduleNames: [
+        "java-debug-parent",
+        "com.microsoft.java.debug.core",
+        "com.microsoft.java.debug.plugin"
+    ]
+};
+
+main(configs, artifactFolder);
+
+function main() {
+    const argv = process.argv;
+    const task = argv[argv.indexOf("-task") + 1];
+    if (task === "upload") {
+        uploadToStaging(configs, artifactFolder);
+    } else if (task === "promote") {
+        promoteToCentral(configs);
+    } else {
+        console.error("Task not specified.");
+        console.log("Usage: node script.js -task [upload|promote]");
+    }
+}
+
+/**
+ * Task upload: Upload artifacts to a nexus staging repo.
  * 
  * Required binaries:
  * - gpg
@@ -13,34 +57,68 @@
  * - NEXUS_STAGINGPROFILEID: identifier of the repo to promote.
  * - GPGPASS: passphrase of GPG key.
  */
-
-const childProcess = require('child_process');
-const fs = require('fs');
-const path = require('path');
-
-const artifactFolder = process.env.artifactFolder;
-const configs = {
-    nexus_ossrhuser: process.env.NEXUS_OSSRHUSER,
-    nexus_ossrhpass: process.env.NEXUS_OSSRHPASS,
-    nexus_stagingProfileId: process.env.NEXUS_STAGINGPROFILEID,
-    gpgpass: process.env.GPGPASS,
-    groupId: "com.microsoft.java",
-    projectName: "java-debug",
-    releaseVersion: process.env.releaseVersion,    
-    moduleNames: [
-        "java-debug-parent",
-        "com.microsoft.java.debug.core",
-        "com.microsoft.java.debug.plugin"
-    ]
-};
-main(configs, artifactFolder);
-
-function main(configs, artifactFolder) {
+function uploadToStaging(configs, artifactFolder) {
     checkPrerequisite(configs);
     addChecksumsAndGpgSignature(configs, artifactFolder);
     createStagingRepo(configs);
     deployToStagingRepo(configs, artifactFolder);
     closeStagingRepo(configs);
+}
+
+ /**
+ * Task promote: Promote a repo to get it picked up by Maven Central.
+ * 
+ * Required binaries:
+ * - curl
+ * 
+ * Required Environment Variables:
+ * - NEXUS_OSSRHUSER: username.
+ * - NEXUS_OSSRHPASS: password.
+ * - NEXUS_STAGINGPROFILEID: identifier of the repo to promote.
+ * - NEXUS_STAGINGREPOID: id of staging repo with artifacts to promote.
+ */
+function promoteToCentral(configs) {
+    let message = "";
+    console.log("\n========Nexus: Promote=======");
+    try {
+        console.log(`Starting to promote staging repository ${configs.stagingRepoId} ...`);
+        console.log(`curl -i -X POST -d "<promoteRequest><data><stagedRepositoryId>${configs.stagingRepoId}</stagedRepositoryId></data></promoteRequest>" -H "Content-Type: application/xml" -u **:** -k https://oss.sonatype.org/service/local/staging/profiles/${configs.nexus_stagingProfileId}/promote`);
+        message = childProcess.execSync(`curl -i -X POST -d "<promoteRequest><data><stagedRepositoryId>${configs.stagingRepoId}</stagedRepositoryId></data></promoteRequest>" -H "Content-Type: application/xml" -u ${configs.nexus_ossrhuser}:${configs.nexus_ossrhpass} -k https://oss.sonatype.org/service/local/staging/profiles/${configs.nexus_stagingProfileId}/promote`);
+        message = message.toString();
+        console.log(message);
+    } catch (ex) {
+        console.error("\n\n[Failure] Promoting staging repository failed.");
+        console.error(!message ? ex : message.toString());
+        process.exit(1);
+    }
+    const success = isReleased(configs);
+    console.log("Below is the public repository url, you could manually validate it.");
+    console.log(`https://oss.sonatype.org/content/groups/public/${configs.groupId.replace(/\./g, "/")}`);
+    console.log("\n\n");
+    if (success) {
+        console.log("\n\n[Success] Nexus: Promote completion.");
+    } else {
+        console.error("\n\n[Failure] Nexus: Promote failed.");
+        process.exit(1)
+    }
+}
+
+function isReleased(configs) {
+    let pollingCount = 0;
+    const MAX_POLLINGS = 10;
+    for (; pollingCount < MAX_POLLINGS; pollingCount++) {
+        console.log(`\nPolling the release operation finished or not...`);
+        console.log(`curl -X GET -H "Content-Type:application/xml" -u **:** -k https://oss.sonatype.org/service/local/staging/repository/${configs.stagingRepoId}`);
+        message = childProcess.execSync(`curl -X GET -H "Content-Type:application/xml" -u ${configs.nexus_ossrhuser}:${configs.nexus_ossrhpass} -k https://oss.sonatype.org/service/local/staging/repository/${configs.stagingRepoId}`);
+        const status = extractStatus(message.toString());
+        console.log(status);
+        if (status !== "closed") {
+            return true;
+        }
+        // use system sleep command to pause the program.
+        childProcess.execSync(`sleep 6s`);
+    }
+    return false;
 }
 
 function checkPrerequisite(configs) {
@@ -135,7 +213,7 @@ function deployToStagingRepo(configs, artifactFolder) {
 function closeStagingRepo(configs) {
     let message = "";
     let pollingCount = 0;
-    const MAX_POLLINGS = 30;
+    const MAX_POLLINGS = 10;
     console.log("\n========Nexus: Verify and Close staging repo=======");
     try {
         console.log(`Starting to close staging repository ${configs.stagingRepoId} ...`);
@@ -152,7 +230,7 @@ function closeStagingRepo(configs) {
                 break;
             }
             // use system sleep command to pause the program.
-            childProcess.execSync(`sleep 2s`);
+            childProcess.execSync(`sleep 6s`);
         }
 
         if (pollingCount >= MAX_POLLINGS) {
@@ -174,32 +252,6 @@ function closeStagingRepo(configs) {
     console.log("\n\n[Success] Nexus: Staging completion.");
     console.log("Below is the staging repository url, you could use it to test deployment.");
     console.log(`https://oss.sonatype.org/content/repositories/${configs.stagingRepoId}`);
-    console.log("\n\n");
-}
-
-function promoteStaging(configs) {
-    let message = "";
-    console.log("\n========Nexus: Promote=======");
-    try {
-        console.log(`Starting to promote staging repository ${configs.stagingRepoId} ...`);
-        console.log(`curl -X POST -d "<promoteRequest><data><stagedRepositoryId>${configs.stagingRepoId}</stagedRepositoryId></data></promoteRequest>" -H "Content-Type: application/xml" -u **:** -k https://oss.sonatype.org/service/local/staging/profiles/${configs.nexus_stagingProfileId}/promote`);
-        message = childProcess.execSync(`curl -X POST -d "<promoteRequest><data><stagedRepositoryId>${configs.stagingRepoId}</stagedRepositoryId></data></promoteRequest>" -H "Content-Type: application/xml" -u ${configs.nexus_ossrhuser}:${configs.nexus_ossrhpass} -k https://oss.sonatype.org/service/local/staging/profiles/${configs.nexus_stagingProfileId}/promote`);
-        message = message.toString();
-        console.log(message);
-        const match = /<stagedRepositoryId>([a-zA-Z0-9-_]+)<\/stagedRepositoryId>/.exec(message);
-        if (match == null || match.length <= 1) {
-            console.error("\n\n[Failure] Promoting staging repository failed.");
-            console.error(message);
-            process.exit(1);
-        }
-    } catch (ex) {
-        console.error("\n\n[Failure] Promoting staging repository failed.");
-        console.error(!message ? ex : message.toString());
-        process.exit(1);
-    }
-    console.log("\n\n[Success] Nexus: Promote completion.");
-    console.log("Below is the public repository url, you could manually validate it.");
-    console.log(`https://oss.sonatype.org/content/groups/public/${configs.groupId.replace(/\./g, "/")}`);
     console.log("\n\n");
 }
